+++ package/Editor/Sources/Services/Packages/Package.cs
@@ -25,11 +25,8 @@ namespace UnityEditor.PackageManager.UI
         }
 
         public PackageInfo Current { get { return Versions.FirstOrDefault(package => package.IsCurrent); } }
-        public PackageInfo Latest { get { return Versions.LastOrDefault(package => package.IsLatest); } }
-        
-        // Package version to display
-        public PackageInfo Display { get { return Current ?? Latest; }}
-
+        public PackageInfo Latest { get { return Versions.LastOrDefault(package => package.IsLatest) ?? Versions.LastOrDefault(); } }
+                
         public IEnumerable<PackageInfo> Versions { get { return source.OrderBy(package => package.Version); } }
         public string Name { get { return packageName; } }